+++ src/Milochau.Emails.Sdk/Models/StorageAccount.cs
@@ -1,9 +1,0 @@
-﻿namespace Milochau.Emails.Sdk.Models
-{
-    /// <summary>Storage account</summary>
-    public enum StorageAccount
-    {
-        /// <summary>Default</summary>
-        Default = 0
-    }
-}