+++ src/Milochau.Emails/DataAccess/StorageDataAccess.cs
@@ -3,7 +3,6 @@ using Azure.Storage.Blobs.Models;
 using Microsoft.Extensions.Options;
 using Milochau.Emails.Sdk.Models;
 using Milochau.Emails.Models.Options;
-using System;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
@@ -21,21 +20,11 @@ namespace Milochau.Emails.DataAccess
 
         public async Task<Stream> ReadToStreamAsync(StorageAttachment attachment, CancellationToken cancellationToken)
         {
-            var connectionString = GetConnectionString(attachment.StorageAccount);
-            var blobServiceClient = new BlobServiceClient(connectionString);
+            var blobServiceClient = new BlobServiceClient(options.DefaultConnectionString);
             var blobContainerClient = blobServiceClient.GetBlobContainerClient(attachment.ContainerName);
             var blobClient = blobContainerClient.GetBlobClient(attachment.FileName);
             
             return await blobClient.OpenReadAsync(new BlobOpenReadOptions(allowModifications: false), cancellationToken);
-        }
-
-        private string GetConnectionString(StorageAccount account)
-        {
-            return account switch
-            {
-                StorageAccount.Default => options.DefaultConnectionString,
-                _ => throw new ArgumentOutOfRangeException(nameof(account), account, $"Unknown storage account {account}")
-            };
         }
     }
 }